feat: added main screen and fixes

+++ lib/services/models/common/forecast_main.dart
@@ -21,15 +21,15 @@ class ForecastMain {
       this.tempKf});
 
   ForecastMain.fromJson(Map<String, dynamic> json) {
-    temp = json['temp'].toDouble();
-    feelsLike = json['feels_like'].toDouble();
-    tempMin = json['temp_min'].toDouble();
-    tempMax = json['temp_max'].toDouble();
+    temp = json['temp'];
+    feelsLike = json['feels_like'];
+    tempMin = json['temp_min'];
+    tempMax = json['temp_max'];
     pressure = json['pressure'];
     seaLevel = json['sea_level'];
     grndLevel = json['grnd_level'];
     humidity = json['humidity'];
-    tempKf = json['temp_kf'].toDouble();
+    tempKf = json['temp_kf'];
   }
 
   Map<String, dynamic> toJson() {
